Introduce base type AbstractBasis for imaginary-time bases.
This removes the real-frequency side as well as the structural dependence
on an SVE-like structure from the basis, since this cannot be reasonably
fulfilled by Legendre, augmented and composite bases.  Focussing on the
imaginary-time side solves this problem.

This also allows us to unify the interfaces of CompositeBasis, LegendreBasis
and ConstBasis (admittedly, a weird case ...)