Remove DimensionlessBasis.
This class was never super-useful for computations, however, it
complicates the basis interface significantly: we always need to
check whether we are dealing with a dimensionless or full basis
when, e.g., composing bases, which was broken anyway.

The real reason for keeping it around was that it provided a
container for holding the SVE, which is now taken care of by
the SVEResult type.